Fix lint, types, tests
