--- src/types/Builder.d.ts
@@ -1,5 +1,0 @@
-import type { IBaseUnit } from './Army'
-
-export interface IBuilderUnit extends IBaseUnit {
-  count: number
-}
