--- src/types/Faction.d.ts
@@ -1,4 +1,0 @@
-export interface IFaction {
-  name: string
-  fileName: string
-}
